requirements.txt
pandas
numpy
scikit-surprise
gradio
scikit-learn
gunicorn
